fix: 删除文件 编辑框依旧可用
