+++ README.md
@@ -15,7 +15,7 @@
 - 回收站功能☑️
 - 数据为空的显示
 - 删除文件不存在报错处理
-- 删除文件 编辑框依旧可用
+- 删除文件 编辑框依旧可用☑️
 
 ## 安装 waild
 ```
